Avoid unnecessary re-renders
This update avoids unnecessary re-renders when the current content of the editor is set again as it's html prop. That can happen when a change propagates through a parent component.

Please take note that this is not only a performance issue, but it also throws the cursor position off if it's at the end of a line.